refactor(sozo): add verbosity flag for `build` cmd
